Make AddRoute + hsts behaviour a shared util
