+++ webapp/routes.go
@@ -6,44 +6,31 @@ package webapp
 
 import (
 	"html/template"
-	"net/http"
+
+	"github.com/web-platform-tests/wpt.fyi/shared"
 )
 
 var templates = template.Must(template.ParseGlob("templates/*.html"))
 
 func init() {
-	routes := map[string]http.HandlerFunc{
-		// Test run results, viewed by browser (default view)
-		// For run results diff view, 'before' and 'after' params can be given.
-		"/": testResultsHandler,
-		"/results": testResultsHandler, // Prevent default redirect
-		"/results/": testResultsHandler,
+	// Test run results, viewed by browser (default view)
+	// For run results diff view, 'before' and 'after' params can be given.
+	shared.AddRoute("/", testResultsHandler)
+	shared.AddRoute("/results", testResultsHandler) // Prevent default redirect
+	shared.AddRoute("/results/", testResultsHandler)
 
-		// About wpt.fyi
-		"/about": aboutHandler,
+	// About wpt.fyi
+	shared.AddRoute("/about", aboutHandler)
 
-		// Test run results, viewed by pass-rate across the browsers
-		"/interop/": interopHandler,
+	// Test run results, viewed by pass-rate across the browsers
+	shared.AddRoute("/interop/", interopHandler)
 
-		// Lists of test run results which have poor interoperability
-		"/interop/anomalies": anomalyHandler,
+	// Lists of test run results which have poor interoperability
+	shared.AddRoute("/interop/anomalies", anomalyHandler)
 
-		// List of all test runs, by SHA[0:10]
-		"/test-runs": testRunsHandler,
+	// List of all test runs, by SHA[0:10]
+	shared.AddRoute("/test-runs", testRunsHandler)
 
-		// Admin-only manual results upload.
-		"/admin/results/upload": adminUploadHandler,
-	}
-
-	for route, handler := range routes {
-		http.HandleFunc(route, wrapHSTS(handler))
-	}
-}
-
-func wrapHSTS(h http.HandlerFunc) http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		value := "max-age=31536000; preload"
-		w.Header().Add("Strict-Transport-Security", value)
-		h(w, r)
-	})
+	// Admin-only manual results upload.
+	shared.AddRoute("/admin/results/upload", adminUploadHandler)
 }
